+++ test/sandbox/app/components/current_user_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class CurrentUserComponent < ViewComponent::Base
+  delegate :current_user, to: :helpers
+end